feat: lift groupByDir state to app level and share across tabs
Move groupByDir toggle state from CommitTab to the top-level App so the
ShelfTab also respects it. Also fix duplicate M-status for newly staged
files and rename the Shelf tab label to Stash.
